FIX ticket #227
**Description**:
 - delete console.logs
 - remove training icon button from home page
 - create training icon in to project-settings

### Related Tickets: #227
